# src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Monash Club Task Manager API",
    "description": "FIT3162 Capstone — Backend API for club task management",
    "version": "1.0.0",
    "contact": { "name": "Team CS-07" }
  },
  "servers": [{ "url": "http://localhost:3000/api/v1", "description": "Local dev" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": { "type": "object" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" }
            }
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "name": { "type": "string" },
          "role": { "type": "string", "enum": ["ADMIN", "MEMBER"] }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string", "enum": ["TODO", "IN_PROGRESS", "DONE"] },
          "priority": { "type": "string", "enum": ["LOW", "MEDIUM", "HIGH"] },
          "assigneeId": { "type": "string", "format": "uuid", "nullable": true },
          "dueDate": { "type": "string", "format": "date-time", "nullable": true },
          "createdBy": { "type": "string", "format": "uuid" },
          "clubId": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "name", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "user@monash.edu" },
                  "name": { "type": "string", "example": "Alice Wong" },
                  "password": { "type": "string", "minLength": 6, "example": "password123" },
                  "role": { "type": "string", "enum": ["ADMIN", "MEMBER"], "default": "MEMBER" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User registered", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "type": "object", "properties": { "user": { "$ref": "#/components/schemas/User" }, "token": { "type": "string" } } } } } } } },
          "409": { "description": "Email already taken" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login and get JWT token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "example": "admin@monash.edu" },
                  "password": { "type": "string", "example": "admin123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "type": "object", "properties": { "user": { "$ref": "#/components/schemas/User" }, "token": { "type": "string" } } } } } } } },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/tasks": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get all tasks (with optional filters)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["TODO", "IN_PROGRESS", "DONE"] } },
          { "name": "priority", "in": "query", "schema": { "type": "string", "enum": ["LOW", "MEDIUM", "HIGH"] } },
          { "name": "assigneeId", "in": "query", "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "List of tasks", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/Task" } } } } } } },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "tags": ["Tasks"],
        "summary": "Create a new task (Admin only)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string", "minLength": 1, "maxLength": 200, "example": "Book venue for O-Week" },
                  "description": { "type": "string", "example": "Find and reserve a venue" },
                  "priority": { "type": "string", "enum": ["LOW", "MEDIUM", "HIGH"], "default": "MEDIUM" },
                  "dueDate": { "type": "string", "format": "date-time" },
                  "assigneeId": { "type": "string", "format": "uuid" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Task created" },
          "403": { "description": "Forbidden (not admin)" }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get a single task by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Task found" },
          "404": { "description": "Task not found" }
        }
      },
      "put": {
        "tags": ["Tasks"],
        "summary": "Update a task (Admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "priority": { "type": "string", "enum": ["LOW", "MEDIUM", "HIGH"] },
                  "status": { "type": "string", "enum": ["TODO", "IN_PROGRESS", "DONE"] },
                  "dueDate": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Task updated" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "tags": ["Tasks"],
        "summary": "Delete a task (Admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "204": { "description": "Task deleted" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/tasks/{id}/assign": {
      "patch": {
        "tags": ["Tasks"],
        "summary": "Assign task to a member (Admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["assigneeId"],
                "properties": {
                  "assigneeId": { "type": "string", "format": "uuid" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Task assigned" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/tasks/{id}/status": {
      "patch": {
        "tags": ["Tasks"],
        "summary": "Change task status (Admin or assigned member)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": { "type": "string", "enum": ["TODO", "IN_PROGRESS", "DONE"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Status changed" },
          "403": { "description": "Forbidden (not admin or assignee)" }
        }
      }
    },
    "/audit": {
      "get": {
        "tags": ["Audit"],
        "summary": "Get audit log (Admin only)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Audit events list" },
          "403": { "description": "Forbidden" }
        }
      }
    }
  }
}